Give up on the anonymous class and make it a normal one to avoid serialization errors
